pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etf-brief"
version = "0.1.0"
description = "Weekly ETF/ETC investment brief with recession signals, macro indicators, and HOLD/INCREASE/DECREASE/SELL recommendations for your saving plans."
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Andrew Buser"}]
requires-python = ">=3.10"
keywords = ["etf", "investing", "recession", "macro", "claude-code-skill"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "pydantic>=2",
]

[project.optional-dependencies]
test = ["pytest>=7", "pytest-mock>=3"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
